+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file, following t
 ## [Unreleased]
 - Mol* core dependency updated to 5.7.0
 - `MVSData.toMVSX` - full support directly from Molstar
+- `.visual.select` function supports `opacity`, `nonSelectedOpacity`, `representationOpacity`
 
 ## [v3.10.1] - 2026-02-05
 - Experimental: Added `MVSData.toMVSX`